refactoring so we get 2 entries in our dropdown menu
