+++ lib/dradis/plugins/burp/html/engine.rb
@@ -1,0 +1,3 @@
+module Dradis::Plugins::Burp::Html
+  class Engine < Dradis::Plugins::Burp::Engine; end
+end
