+++ lib/dradis/plugins/burp/xml/engine.rb
@@ -1,0 +1,3 @@
+module Dradis::Plugins::Burp::Xml
+  class Engine < Dradis::Plugins::Burp::Engine; end
+end
